Rename `Login` component to `SignIn`
The qualifying tests app has used the 'sign in' (rather than 'login') terminology. However one legacy remained: the sign in page component was called `Login`. For consistency, I've renamed it to `SignIn`.

This mirrors some upcoming changes to the application form app which will make the same move away from 'login' in favour of 'sign in'.
